cheri: Fix compile errors in realtek wirless drivers
- Change the layout of struct ieee80211_tx_info. We need more
  space as pointers are larger on CHERI. The structure is embedded
  in opaque data of an sk_buff that we increase for CHERI, too.
  To make sure that everything fits, add a static_assert() that
  checks the size.
- Fix a address vs. value confusion in upstream code that triggers
  and error with CHERI.
- Fix sparse warnings and compile errors.

+++ drivers/net/wireless/realtek/rtlwifi/core.c
@@ -774,7 +774,7 @@ static int rtl_op_config(struct ieee80211_hw *hw, u32 changed)
 
 static void rtl_op_configure_filter(struct ieee80211_hw *hw,
 				    unsigned int changed_flags,
-				    unsigned int *new_flags, u64 multicast)
+				    unsigned int *new_flags, uintptr_t multicast)
 {
 	bool update_rcr = false;
 	struct rtl_priv *rtlpriv = rtl_priv(hw);
